move template into cpp
